chore: update wasm-js to wrap-js
