--- packages/js/wasm/src/IWasmPackage.ts
@@ -1,6 +1,0 @@
-import { IWrapPackage } from "@polywrap/core-js";
-import { Result } from "@polywrap/result";
-
-export interface IWasmPackage extends IWrapPackage {
-  getWasmModule(): Promise<Result<Uint8Array, Error>>;
-}
