+++ packages/js/wasm/src/helpers/types.ts
@@ -1,0 +1,1 @@
+export type u32 = number;
