+++ packages/js/wasm/src/index.ts
@@ -1,7 +1,7 @@
-export { WasmWrapper } from "./WasmWrapper";
-export { WasmPackage } from "./WasmPackage";
-export { IWasmPackage } from "./IWasmPackage";
-export { IFileReader } from "./IFileReader";
-export { InMemoryFileReader } from "./InMemoryFileReader";
-export { WrapImports } from "./types";
 export * from "./constants";
+export * from "./WasmWrapper";
+export * from "./WasmPackage";
+export * from "./IFileReader";
+export * from "./InMemoryFileReader";
+export * from "./WrapImports";
+export * from "./WrapExports";
