+++ packages/js/wrap/src/invoke/Invoker.ts
@@ -1,5 +1,9 @@
 import { InvokeOptions, InvokeResult } from ".";
+import { Uri, UriResolutionContext } from "../uri-resolution";
 import { Wrapper } from "../wrapper";
+import { WrapError } from "../WrapError";
+
+import { Result } from "@polywrap/result";
 
 /**
  * Provides options for the invoker to set based on the state of the invocation.
@@ -8,6 +12,15 @@ import { Wrapper } from "../wrapper";
 export interface InvokerOptions extends InvokeOptions {
   /** If true, the InvokeResult will (if successful) contain a Msgpack-encoded byte array */
   encodeResult?: boolean;
+}
+
+/** Options for CoreClient's getImplementations method */
+export interface GetImplementationsOptions {
+  /** If true, follow redirects to resolve URIs */
+  applyResolution?: boolean;
+
+  /** Use and update an existing resolution context */
+  resolutionContext?: UriResolutionContext;
 }
 
 /**
@@ -38,4 +51,17 @@ export interface Invoker {
   invoke<TData = unknown>(
     options: InvokerOptions
   ): Promise<InvokeResult<TData>>;
+
+  /**
+   * returns the interface implementations associated with an interface URI
+   *  from the configuration used to instantiate the client
+   *
+   * @param uri - a wrap URI
+   * @param options - { applyResolution?: boolean; resolutionContext?: IUriResolutionContext }
+   * @returns a Result containing URI array if the request was successful
+   */
+  getImplementations(
+    uri: Uri,
+    options: GetImplementationsOptions
+  ): Promise<Result<Uri[], WrapError>>;
 }
